server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
fastapi==0.73.0
guess-language-spirit==0.5.3
h11==0.13.0
httptools==0.3.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.2
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.0
pydantic==1.9.0
pypinyin==0.46.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
watchgod==0.7
websockets==10.1
